+++ NEWS.md
@@ -1,4 +1,4 @@
-# thurstonianIRT 0.10.0++
+# thurstonianIRT 0.11.0
 
 ## Bug Fixes
 
@@ -9,9 +9,10 @@ thanks to @IanEisenberg. (#13)
 
 ## New Features
 
-* Support predictions of trait scores for new persons. (#12)
+* Support predictions of trait scores for new persons. (#12, #15)
 * Specify multiple blocks at once via `set_blocks_from_df`
 thanks to @awhug. (#11)
+
 
 # thurstonianIRT 0.10.0